refactor(protocol): reorganize flat files into domain-based directory structure
